// node/typings/Order.ts
export type Order = {
  orderId: string
  creationDate: Date
  clientName: string
  items?: any
  totalValue: number
  paymentNames: string
  status: string
  statusDescription: string
  marketPlaceOrderId?: any
  sequence: string
  salesChannel: string
  affiliateId: string
  origin: string
  workflowInErrorState: boolean
  workflowInRetry: boolean
  lastMessageUnread: string
  ShippingEstimatedDate?: any
  ShippingEstimatedDateMax?: any
  ShippingEstimatedDateMin?: any
  orderIsComplete: boolean
  listId?: any
  listType?: any
  authorizedDate?: any
  callCenterOperatorName?: any
  totalItems: number
  currencyCode: string
  hostname: string
  invoiceOutput?: any
  invoiceInput?: any
  lastChange: Date
  isAllDelivered: boolean
  isAnyDelivered: boolean
  giftCardProviders?: any
  orderFormId: string
  paymentApprovedDate?: any
  readyForHandlingDate?: any
  deliveryDates?: any
}